refactor(intel-dcap-api): split client.rs into smaller files
